custom_ordenes_v2/requirements.txt
Flask
requests
python-dotenv
azure-cognitiveservices-vision-customvision
msrest
Flask-SQLAlchemy
Flask-JWT-Extended
openai